refactor: centralize search interactions
